[bug/VSPC-207] changed the code based on the review comment
